Organise all code into different files, set up API call to search locations
