+++ src/javascript/startApp.js
@@ -1,0 +1,7 @@
+import setSearchBtn from './eventListeners.js';
+
+const start = () => {
+  setSearchBtn();
+};
+
+export default start;
